creating , union, intersect, diff & printing sets
exercise guidelines:
Define a function singletonSet which creates a singleton set from one integer value: the set represents the set of the one given element. Now that we have a way to create singleton sets, we want to define a function that allow us to build bigger sets from smaller ones.
Define the functions union,intersect, and diff, which takes two sets, and return, respectively, their union, intersection and differences. diff(s, t) returns a set which contains all the elements of the set s that are not in the set t.
Define the function filter which selects only the elements of a set that are accepted by a given predicate p. The filtered elements are returned as a new set.